+++ UnitTests/Specification/AllAttributesTest.php
@@ -4,16 +4,16 @@ require_once dirname(__FILE__)
 	. DIRECTORY_SEPARATOR . '..'
 	. DIRECTORY_SEPARATOR . 'TestHelper.php';
 
-class ElementTree_Filter_AllElementsTest extends PHPUnit_Framework_TestCase
+class ElementTree_Specification_AllAttributesTest extends PHPUnit_Framework_TestCase
 {
 	/**
 	 * @test
 	 */
-	public function isSatisfiedWhenGivenElementComponent()
+	public function isSatisfiedWhenGivenAttributeComponent()
 	{
-		$allText = new \ElementTree\Filter\AllElements();
+		$allText = new \ElementTree\Specification\AllAttributes();
 		$this->assertTrue(
-			$allText->isSatisfiedBy(new \ElementTree\ElementTreeElement('a'))
+			$allText->isSatisfiedBy(new \ElementTree\ElementTreeAttribute('foo', 'bar'))
 		);
 	}
 
@@ -22,7 +22,7 @@ class ElementTree_Filter_AllElementsTest extends PHPUnit_Framework_TestCase
 	 */
 	public function isNotSatisfiedWhenGivenElementComponent()
 	{
-		$allText = new \ElementTree\Filter\AllElements();
+		$allText = new \ElementTree\Specification\AllAttributes();
 		$this->assertFalse(
 			$allText->isSatisfiedBy(new \ElementTree\ElementTreeText('a'))
 		);