+++ UnitTests/Specification/AllTextTest.php
@@ -4,14 +4,14 @@ require_once dirname(__FILE__)
 	. DIRECTORY_SEPARATOR . '..'
 	. DIRECTORY_SEPARATOR . 'TestHelper.php';
 
-class ElementTree_Filter_AllTextTest extends PHPUnit_Framework_TestCase
+class ElementTree_Specification_AllTextTest extends PHPUnit_Framework_TestCase
 {
 	/**
 	 * @test
 	 */
 	public function isSatisfiedWhenGivenTextComponent()
 	{
-		$allText = new \ElementTree\Filter\AllText();
+		$allText = new \ElementTree\Specification\AllText();
 		$this->assertTrue(
 			$allText->isSatisfiedBy(new \ElementTree\ElementTreeText('a'))
 		);
@@ -22,7 +22,7 @@ class ElementTree_Filter_AllTextTest extends PHPUnit_Framework_TestCase
 	 */
 	public function isNotSatisfiedWhenGivenElementComponent()
 	{
-		$allText = new \ElementTree\Filter\AllText();
+		$allText = new \ElementTree\Specification\AllText();
 		$this->assertFalse(
 			$allText->isSatisfiedBy(new \ElementTree\ElementTreeElement('a'))
 		);