+++ UnitTests/Specification/AndSpecificationTest.php
@@ -4,16 +4,16 @@ require_once dirname(__FILE__)
 	. DIRECTORY_SEPARATOR . '..'
 	. DIRECTORY_SEPARATOR . 'TestHelper.php';
 
-class ElementTree_Filter_AndSpecificationTest extends PHPUnit_Framework_TestCase
+class ElementTree_Specification_AndSpecificationTest extends PHPUnit_Framework_TestCase
 {
 	/**
 	 * @test
 	 */
 	public function isSatisfiedWhenAllSubspecificationsAreSatisfied()
 	{
-		$andSpec = new \ElementTree\Filter\AndSpecification(
-			new \ElementTree\Filter\ElementByName('a'),
-			new \ElementTree\Filter\ElementByName('a')
+		$andSpec = new \ElementTree\Specification\AndSpecification(
+			new \ElementTree\Specification\AllElements(),
+			new \ElementTree\Specification\WithName('a')
 		);
 
 		$this->assertTrue($andSpec->isSatisfiedBy(new \ElementTree\ElementTreeElement('a')));
@@ -24,9 +24,9 @@ class ElementTree_Filter_AndSpecificationTest extends PHPUnit_Framework_TestCase
 	 */
 	public function isNotSatisfiedWhenNotAllSubspecificationsAreSatisfied()
 	{
-		$andSpec = new \ElementTree\Filter\AndSpecification(
-			new \ElementTree\Filter\ElementByName('a'),
-			new \ElementTree\Filter\ElementByName('b')
+		$andSpec = new \ElementTree\Specification\AndSpecification(
+			new \ElementTree\Specification\AllElements(),
+			new \ElementTree\Specification\WithName('b')
 		);
 
 		$this->assertFalse($andSpec->isSatisfiedBy(new \ElementTree\ElementTreeElement('a')));