+++ UnitTests/Specification/NotSpecificationTest.php
@@ -4,15 +4,15 @@ require_once dirname(__FILE__)
 	. DIRECTORY_SEPARATOR . '..'
 	. DIRECTORY_SEPARATOR . 'TestHelper.php';
 
-class ElementTree_Filter_NotSpecificationTest extends PHPUnit_Framework_TestCase
+class ElementTree_Specification_NotSpecificationTest extends PHPUnit_Framework_TestCase
 {
 	/**
 	 * @test
 	 */
 	public function isSatisfiedWhenSubspecificationIsNotSatisfied()
 	{
-		$notSpec = new \ElementTree\Filter\NotSpecification(
-			new \ElementTree\Filter\ElementByName('a')
+		$notSpec = new \ElementTree\Specification\NotSpecification(
+			new \ElementTree\Specification\WithName('a')
 		);
 
 		$this->assertTrue($notSpec->isSatisfiedBy(new \ElementTree\ElementTreeElement('b')));
@@ -23,8 +23,8 @@ class ElementTree_Filter_NotSpecificationTest extends PHPUnit_Framework_TestCase
 	 */
 	public function isNotSatisfiedWhenSubspecificationIsSatisfied()
 	{
-		$notSpec = new \ElementTree\Filter\NotSpecification(
-			new \ElementTree\Filter\ElementByName('a')
+		$notSpec = new \ElementTree\Specification\NotSpecification(
+			new \ElementTree\Specification\WithName('a')
 		);
 
 		$this->assertFalse($notSpec->isSatisfiedBy(new \ElementTree\ElementTreeElement('a')));