+++ UnitTests/Specification/OrSpecificationTest.php
@@ -4,16 +4,16 @@ require_once dirname(__FILE__)
 	. DIRECTORY_SEPARATOR . '..'
 	. DIRECTORY_SEPARATOR . 'TestHelper.php';
 
-class ElementTree_Filter_OrSpecificationTest extends PHPUnit_Framework_TestCase
+class ElementTree_Specification_OrSpecificationTest extends PHPUnit_Framework_TestCase
 {
 	/**
 	 * @test
 	 */
 	public function isSatisfiedWhenAtLeastOneSubspecificationsIsSatisfied()
 	{
-		$orSpec = new \ElementTree\Filter\OrSpecification(
-			new \ElementTree\Filter\ElementByName('a'),
-			new \ElementTree\Filter\ElementByName('b')
+		$orSpec = new \ElementTree\Specification\OrSpecification(
+			new \ElementTree\Specification\WithName('a'),
+			new \ElementTree\Specification\WithName('b')
 		);
 
 		$this->assertTrue($orSpec->isSatisfiedBy(new \ElementTree\ElementTreeElement('a')));
@@ -24,9 +24,9 @@ class ElementTree_Filter_OrSpecificationTest extends PHPUnit_Framework_TestCase
 	 */
 	public function isNotSatisfiedWhenAllSubspecificationsAreNotSatisfied()
 	{
-		$orSpec = new \ElementTree\Filter\OrSpecification(
-			new \ElementTree\Filter\ElementByName('b'),
-			new \ElementTree\Filter\ElementByName('c')
+		$orSpec = new \ElementTree\Specification\OrSpecification(
+			new \ElementTree\Specification\WithName('b'),
+			new \ElementTree\Specification\WithName('c')
 		);
 
 		$this->assertFalse($orSpec->isSatisfiedBy(new \ElementTree\ElementTreeElement('a')));